fix color
